--- routes/restorePasswordApiRoutes.js
@@ -1,8 +1,0 @@
-const express = require("express")
-const router = express.Router()
-const passwordController = require("../controllers/passwordControllers")
-
-// [GET] /restorepassword Cambiar password
-router.get("/",passwordController.restorePassword) 
-
-module.exports = router
